New branch

+++ etc/DockerHelper.sh
@@ -33,6 +33,7 @@ usage: $0 [CMD] [OPTIONS]
   -password=PASSWORD            Password to loging at the docker registry.
   -ci                           Install CI tools in image
   -dry-run                      Do not push images to the repository
+  -push-latest                  Push the latest image to the repository
   -no-constant-build-dir        Do not use constant build directory
   -h -help                      Show this message and exits
 
@@ -64,12 +65,18 @@ _setup() {
             fromImage="${FROM_IMAGE_OVERRIDE:-"${org}/flow-${os}-dev"}:${imageTag}"
             context="."
             buildArgs="--build-arg numThreads=${numThreads}"
+            orVersion=$(git -C tools/OpenROAD describe --tags)
+            echo "OpenROAD version: ${orVersion}"
+            buildArgs+=" --build-arg openroadVersion=${orVersion}"
             ;;
         "dev" )
             fromImage="${FROM_IMAGE_OVERRIDE:-$osBaseImage}"
             cp tools/OpenROAD/etc/DependencyInstaller.sh etc/InstallerOpenROAD.sh
             context="etc"
-            buildArgs="--build-arg options=${options} ${noConstantBuildDir}"
+            local yosys_ver
+            yosys_ver=v$(grep 'yosys_ver =' tools/yosys/docs/source/conf.py | awk -F'"' '{print $2}')
+            options+=" -yosys-ver=${yosys_ver}"
+            buildArgs="--build-arg \"options=${options}\" ${noConstantBuildDir}"
             ;;
         *)
             echo "Target ${target} not found" >&2
@@ -83,20 +90,20 @@ _setup() {
 
 _create() {
     echo "Create docker image ${imagePath} using ${file}"
-    ${DOCKER_CMD} buildx build \
+    eval ${DOCKER_CMD} buildx build \
         --file "${file}" \
         --tag "${imagePath}" \
-        ${buildArgs} \
+        "${buildArgs}" \
         "${context}"
     rm -f etc/InstallerOpenROAD.sh
 }
 
 _push() {
-    if [[ -z ${username+x} ]]; then
+    if [[ -z ${username+x} ]] && [[ ${dryRun} != 1 ]]; then
         echo "Missing required -username=<USER> argument"
         _help
     fi
-    if [[ -z ${password+x} ]]; then
+    if [[ -z ${password+x} ]] && [[ ${dryRun} != 1 ]]; then
         echo "Missing required -password=<PASS> argument"
         _help
     fi
@@ -136,10 +143,22 @@ _push() {
         orfsTag=${org}/orfs:${tag}
         echo "Renaming docker image: ${builderTag} -> ${orfsTag}"
         ${DOCKER_CMD} tag ${builderTag} ${orfsTag}
+
         if [[ "${dryRun}" == 1 ]]; then
             echo "[DRY-RUN] ${DOCKER_CMD} push ${orfsTag}"
+            if [[ "${pushLatest}" == 1 ]]; then
+                echo "[DRY-RUN] ${DOCKER_CMD} tag ${orfsTag} \"${org}/orfs:latest\""
+                echo "[DRY-RUN] ${DOCKER_CMD} push \"${org}/orfs:latest\""
+            fi
         else
             ${DOCKER_CMD} push ${orfsTag}
+
+            # Only tag and push as latest if requested
+            if [[ "${pushLatest}" == 1 ]]; then
+                ${DOCKER_CMD} tag ${orfsTag} "${org}/orfs:latest"
+                ${DOCKER_CMD} push "${org}/orfs:latest"
+                echo "Tagged and pushed ${org}/orfs:latest"
+            fi
         fi
     fi
 }
@@ -174,6 +193,7 @@ numThreads="-1"
 tag=""
 options=""
 dryRun=0
+pushLatest=0
 
 while [ "$#" -gt 0 ]; do
     case "${1}" in
@@ -185,6 +205,9 @@ while [ "$#" -gt 0 ]; do
             ;;
         -dry-run )
             dryRun=1
+            ;;
+        -push-latest )
+            pushLatest=1
             ;;
         -os=* )
             os="${1#*=}"
